# apps/backend/src/routes/transaction/sync-coingate.ts
import { FastifyInstance } from "fastify"
import { eq } from "drizzle-orm"
import { db } from "../../db.js"
import { verifyAuth } from "../../middleware/auth.js"
import { transactions } from "../../schema.js"
import type { GenericRouteRequest, RouteReply } from "../../types/requests.js"
import { normalizeCoinGateOrderStatus } from "../../services/coingate-payments.js"
import { ingestPaymentWebhook } from "../../services/payment-webhooks.js"
import { getErrorMessage } from "../../utils/error.js"
import { readStringMetadata, transactionParamsSchema } from "./shared.js"
import { syncCoinGateSchema } from "../../schemas/transaction.schemas.js"

export async function registerTransactionSyncRoutes(server: FastifyInstance) {
  server.post(
    "/transactions/:transactionId/sync-coingate",
    { preHandler: verifyAuth, schema: syncCoinGateSchema },
    async (request: GenericRouteRequest, reply: RouteReply) => {
      try {
        const { transactionId } = transactionParamsSchema.parse(request.params)
        const userId = request.user!.userId

        const [transaction] = await db
          .select({
            id: transactions.id,
            userId: transactions.userId,
            method: transactions.method,
            status: transactions.status,
            metadata: transactions.metadata,
          })
          .from(transactions)
          .where(eq(transactions.id, transactionId))
          .limit(1)

        if (!transaction) {
          return reply.status(404).send({ error: "Transaction not found" })
        }

        if (transaction.userId !== userId) {
          return reply.status(403).send({ error: "Forbidden" })
        }

        if (transaction.method !== "coingate_checkout") {
          return reply.status(400).send({
            error: "Transaction is not a CoinGate checkout",
          })
        }

        const coingateOrderId = readStringMetadata(
          transaction.metadata,
          "coingate_order_id"
        )

        if (!coingateOrderId) {
          return reply.status(400).send({
            error: "Transaction is missing CoinGate order metadata",
          })
        }

        const parsedOrderId = Number.parseInt(coingateOrderId, 10)

        if (!Number.isFinite(parsedOrderId)) {
          return reply.status(400).send({
            error: "Transaction has invalid CoinGate order metadata",
          })
        }

        const payload = await normalizeCoinGateOrderStatus(parsedOrderId)
        const result = await ingestPaymentWebhook({
          provider: "coingate",
          payload,
        })

        return {
          message: "CoinGate transaction sync complete",
          ...result,
        }
      } catch (err: unknown) {
        request.log.error(err)
        return reply.status(400).send({
          error: getErrorMessage(err, "Unable to sync CoinGate transaction"),
        })
      }
    }
  )
}